refactoring to repository + handlers for aggregate
