+++ koopovereenkomst-v3-simple/pages/steps_verkoper/Step5.tsx
@@ -7,12 +7,12 @@ import { getInteger, getSolidDataset, getThingAll } from '@inrupt/solid-client';
 import { fetch } from '@inrupt/solid-client-authn-browser';
 import { useSession } from "@inrupt/solid-ui-react";
 import { useCallback, useEffect, useState } from "react";
-import { SOLID_ZVG_CONTEXT } from '../../src/aggregate/context';
+import { SOLID_ZVG_CONTEXT } from '../../src/koek/Context';
 
 import { List, ListItem, ListItemText, TextField } from "@mui/material";
 import { default as data } from "@solid/query-ldflex/lib/exports/rdflib";
 import Link from "../../src/Link";
-import KoopovereenkomstAggregate from '../../src/aggregate/koopovereenkomst-aggregate';
+import KoekAggregate from '../../src/koek/KoekAggregate';
 
 const zvg_base = 'http://taxonomie.zorgeloosvastgoed.nl/def/zvg#'
 const zvg = {
@@ -28,7 +28,7 @@ interface Step5Props {
   stepNr: number;
   handleNext: () => void;
   handleBack: () => void;
-  koek: KoopovereenkomstAggregate;
+  koek: KoekAggregate;
 }
 
 export function Step5({ stepNr = 5, handleNext, handleBack, koek }: Step5Props) {
@@ -55,16 +55,11 @@ export function Step5({ stepNr = 5, handleNext, handleBack, koek }: Step5Props) 
 
   const openenKoopovereenkomstWithInruptSolidClient = async function () {
     try {
-      setCase({});
+      console.log(`Openen van Koopovereenkomst [caseId: ${caseId}]`);
 
       const theKO = await getSolidDataset(koopovereenkomstFile(), {
         fetch: fetch,
       });
-
-      setCase((prevState) => ({
-        ...prevState,
-        caseId: caseId,
-      }));
 
       for (const thing of getThingAll(theKO)) {
         // console.log(JSON.stringify(thing));
@@ -92,23 +87,24 @@ export function Step5({ stepNr = 5, handleNext, handleBack, koek }: Step5Props) 
       const pod = webId.split("/profile/card#me")[0];
       // TODO: this only works for public data
       // for private data we need to use the authenticated fetch
-      const ko = data[`${pod}/koopovereenkomst/id/${caseId}`];
+      // const ko = data[`${pod}/koopovereenkomst/id/${caseId}`];
 
-      const aggregate = new KoopovereenkomstAggregate(ko, caseId);
+      // const aggregate = new KoekAggregate(ko, caseId);
+      const aggregate = koek;
 
-      try {
-        console.log("\nGENERATED BY");
-        for await (const eventUri of ko.wasGeneratedBy) {
-          await aggregate.handleEvent(eventUri);
-        }
+      // try {
+      //   console.log("\nGENERATED BY");
+      //   for await (const eventUri of ko.wasGeneratedBy) {
+      //     await aggregate.handleEvent(eventUri);
+      //   }
 
         setEventLabels(aggregate.getEvents());
 
-        console.log(aggregate.getEvents());
-      } catch (error) {
-        setErrors("Error handling events of koopovereenkomst! " + error);
-        console.error("Error handling events of koopovereenkomst!", error);
-      }
+      //   console.log(aggregate.getEvents());
+      // } catch (error) {
+      //   setErrors("Error handling events of koopovereenkomst! " + error);
+      //   console.error("Error handling events of koopovereenkomst!", error);
+      // }
 
       console.log(
         `\nAANGEBODEN DOOR: ${await aggregate.getData().aangebodenDoor}`
