+++ koopovereenkomst-v3-simple/src/kkg/Perceel.ts
@@ -1,0 +1,4 @@
+interface Perceel {
+  kadastraalObjectId: string;
+  perceelNummer: string;
+}
