+++ koopovereenkomst-v3-simple/src/rdfWriter.ts
@@ -1,6 +1,6 @@
 import * as $rdf from 'rdflib';
 
-import { Event } from './aggregate/koopovereenkomst-aggregate';
+import { Event } from "./koek/Event";
 import { getWebId } from './Solid';
 
 export function Aggregate2RDF(id: string, events: Event[], options): string {
@@ -20,6 +20,7 @@ export function Aggregate2RDF(id: string, events: Event[], options): string {
 
   const eventList = events.map(event => { 
     const eventContainer = event.actor === "koper-koos" ? `http://localhost:3001/koper-koos/koopovereenkomst/events/id` : options.eventContainer;
+    console.log(`event.id [${event.id}] converted to `, eventContainer);
     return `${eventContainer}/${event.id}`
   });
